requirements.txt
Django==3.2.20
djangorestframework==3.11.0
django-cors-headers==3.2.1
djongo==1.3.6
pymongo==3.12.3
requests=2.31.0
